Nuke onnx models from test  (#6890)
* Elimitated usage of onnx models in tests/ disabled some tests

* Verified green: batch 1

* Verified tests: batch 2

* Verified tests: batch 3

* Verified tests: batch 4

* Verified tests: batch 5

* Verified tests: batch 6

* Verified tests: batch 7

* Verified tests: batch 8

* Verified tests: batch 9

* Verified tests: batch 10

* Verified tests: batch 11

* Verified tests: batch 12

* Verified tests: batch 12_follow up

* Verified tests: batch 13

* Verified tests: batch 14

* Verified tests: batch 15

* Verified tests: batch 16

* Removed exits

+++ qa/L0_batch_input/test.sh
@@ -60,6 +60,9 @@ rm -f $SERVER_LOG $CLIENT_LOG
 
 RET=0
 for BACKEND in $BACKENDS; do
+    if [[ "$BACKEND" == 'onnx' ]]; then
+        continue
+    fi
     rm -rf models && mkdir models
     cp -r $DATADIR/${BACKEND}_batch_input models/ragged_element_count_acc_zero
     (cd models/ragged_element_count_acc_zero && \
